refactor(namespace): move spec loading components to Spec/
Move 7 spec-loading classes out of the top-level namespace into a dedicated
Spec/ subnamespace. This signals that they form a cohesive unit and pre-takes
the boundary that a future package extraction would draw — without requiring
a monorepo split.

Moved:
- Studio\OpenApiContractTesting\OpenApiSpecLoader      -> Spec\OpenApiSpecLoader
- Studio\OpenApiContractTesting\OpenApiSpecResolver    -> Spec\OpenApiSpecResolver
- Studio\OpenApiContractTesting\OpenApiRefResolver     -> Spec\OpenApiRefResolver
- Studio\OpenApiContractTesting\RefResolutionContext   -> Spec\RefResolutionContext
- Studio\OpenApiContractTesting\OpenApiSchemaConverter -> Spec\OpenApiSchemaConverter
- Studio\OpenApiContractTesting\OpenApiPathMatcher     -> Spec\OpenApiPathMatcher
- Studio\OpenApiContractTesting\OpenApiPathSuggester   -> Spec\OpenApiPathSuggester

Top-level callers (OpenApiResponseValidator, OpenApiRequestValidator,
OpenApiCoverageTracker) now import the relocated classes explicitly.
OpenApiSchemaConverter, now in Spec/, imports the still-top-level
OpenApiVersion and SchemaContext enums.

BREAKING: Pre-1.0 namespace reorg, allowed by CHANGELOG. v0.15.0 target.

PHPUnit 1022/1022 passing. PHPStan level 6 clean. PHP-CS-Fixer clean.
