+++ src/Spec/OpenApiSpecResolver.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Studio\OpenApiContractTesting;
+namespace Studio\OpenApiContractTesting\Spec;
 
 use ReflectionClass;
 use ReflectionMethod;
