+++ tests/Unit/CoverageReportSubscriberWorkerModeTest.php
@@ -9,10 +9,10 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use Studio\OpenApiContractTesting\OpenApiCoverageTracker;
-use Studio\OpenApiContractTesting\OpenApiSpecLoader;
 use Studio\OpenApiContractTesting\PHPUnit\ConsoleOutput;
 use Studio\OpenApiContractTesting\PHPUnit\CoverageReportSubscriber;
 use Studio\OpenApiContractTesting\PHPUnit\CoverageSidecarReader;
+use Studio\OpenApiContractTesting\Spec\OpenApiSpecLoader;
 use Throwable;
 
 use function file_get_contents;
