+++ tests/Unit/Laravel/ExploresOpenApiEndpointTest.php
@@ -12,7 +12,7 @@ use Studio\OpenApiContractTesting\Fuzz\ExplorationCases;
 use Studio\OpenApiContractTesting\Fuzz\ExploredCase;
 use Studio\OpenApiContractTesting\HttpMethod;
 use Studio\OpenApiContractTesting\Laravel\ExploresOpenApiEndpoint;
-use Studio\OpenApiContractTesting\OpenApiSpecLoader;
+use Studio\OpenApiContractTesting\Spec\OpenApiSpecLoader;
 
 // Re-use the namespace-level config() mock so config()-based fallback works.
 require_once __DIR__ . '/../../Helpers/LaravelConfigMock.php';
